changed database value to float

+++ Project-Server/MySQL/tempHistory.sql
@@ -1,6 +1,6 @@
 CREATE TABLE TempHistory(
     tempID INT NOT NULL AUTO_INCREMENT,
-    tempC DECIMAL(5,2) NOT NULL,
+    tempC FLOAT NOT NULL,
     tempDate DATE NOT NULL,
     tempTime TIME NOT NULL,
     PRIMARY KEY (tempID)
